MA-7485 add VideoStop.ogg in AllAudio.mk
In libcameraservice, mm6.0 uses VideoStop.ogg as the video record stop sound.
This ogg file is located at frameworks/base/data/sounds/effects/ogg/,
but it has not been added to the AllAudio.mk which is included in our imx6/7.mk.

Add this file in the AllAudio.mk to fix this issue.

Signed-off-by: Winter Wang <b55847@freescale.com>